refactor: reorganize image and signatures validation modules for improved structure and clarity
