--- docs/api/entities/helpers.md
@@ -1,1 +1,0 @@
-::: siapy.entities.helpers
