--- docs/api/entities/images/validators.md
@@ -1,1 +1,0 @@
-::: siapy.entities.images.validators
