+++ mkdocs.yml
@@ -89,12 +89,10 @@ nav:
               - Spectral Library: api/entities/images/spectral_lib.md
               - Mock Image: api/entities/images/mock.md
               - Spectral Images: api/entities/images/spimage.md
-              - Validators: api/entities/images/validators.md
           - Shape: api/entities/shapes/shape.md
           - Image Sets: api/entities/imagesets.md
           - Pixels: api/entities/pixels.md
           - Signatures: api/entities/signatures.md
-          - Helpers: api/entities/helpers.md
       - Features:
           - Features: api/features/features.md
           - Helpers: api/features/helpers.md
@@ -111,7 +109,9 @@ nav:
           - Image: api/transformations/image.md
       - Utils:
           - Images: api/utils/images.md
+          - Images Validators: api/utils/image_validators.md
           - Plots: api/utils/plots.md
+          - Signatures: api/utils/signatures.md
   - Release Notes: changelog.md
   - License: permit.md
