+++ siapy/utils/signatures.py
